Added reproject option to MODIS processing.
Added GDAL_dir to calc_average to support OS processing. If not specified in ‘process', it is required under ‘directory’ in the config.

Added GDAL_dir to VCI, TCI and VHI to support OS processing. If not specified in ‘process', it is required under ‘directory’ in the config.

Pass toolDir into calcVCI_os, calcTCI_os and calcVHI_os

Added overwrite and no_nodata options to crop raster.

Added raster resample function.